Add layout and page components in [lang]
